feat: forward to loos page with message after contribution
